Simplify down the project & refactor
